ui: charts cleanup
